requirements.txt
accelerate==0.15.0
certifi==2023.7.22
charset-normalizer==3.2.0
clean-fid==0.1.35
diffusers==0.12.0
filelock==3.12.4
fsspec==2023.9.1
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
lpips==0.1.4
numpy==1.26.0
opencv-python==4.7.0.68
packaging==23.1
pillow==10.0.1
psutil==5.9.5
pyyaml==6.0.1
regex==2023.8.8
requests==2.31.0
scipy==1.11.2
tokenizers==0.13.3
torch-fidelity==0.3.0
torchmetrics==0.11.0
tqdm==4.66.1
transformers==4.25.1
typing-extensions==4.8.0
urllib3==2.0.5
zipp==3.17.0
